requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
fanstatic==1.5
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
js.fullcalendar==2.9.1
js.jquery==3.6.1
js.momentjs==2.13.1
legacy-cgi==2.6.3
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyDrive==1.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
reportlab==4.4.1
setuptools==80.6.0
six==1.17.0
tzdata==2025.2
WebOb==1.8.9
Werkzeug==3.1.3
WTForms==3.2.1
supabase==2.17.0